166 native elements (#174)
* removing unnecessary functionality from buton, applying it to the link

* correcting resource loader for linkButtons
