+++ src/components/BioLinkButton/BioLinkButton.ts
@@ -12,6 +12,26 @@ class BioLinkButton extends Component< BioLinkButtonProps, BioLinkButtonState > 
 
 	public methods: BioLinkButtonMethods = {};
 
+	connectedCallback() {
+		this.registerEventListeners();
+	}
+
+	registerEventListeners() {
+		if (this.props.url) {
+			this.addEventListener('click', (e: Event) => {
+				e.preventDefault();
+				if (this.props.url[0] === '#') {
+					const targetElement: HTMLElement = document.querySelector(this.props.url);
+					if (!!targetElement) {
+						targetElement.scrollIntoView({behavior: 'smooth'});
+					}
+				} else {
+					window.open(this.props.url);
+				}
+			});
+		}
+	}
+
 	get defaultState() {
 		return {};
 	}
