+++ src/components/BioLinkButton/styles.scss
@@ -5,8 +5,9 @@
     a {
 		$border: 2px;
 
-		min-width: 160px;
-		padding: 10px 42px;
+		display: block;
+		width: 160px;
+		padding: 10px 0px;
 		background: transparent;
 		color: $bio-black;
 		border: $border solid $bio-orange;
@@ -24,6 +25,32 @@
 			border: 2px solid transparent;
 			background: $bio-orange;
 			color: $bio-white;
+		}
+
+        &.blue {
+			border: $border solid $bio-dark-blue;
+			
+			&:hover{
+				background: $bio-dark-blue;
+			}
+		}
+		
+		&.white {
+			border: $border solid $bio-white;
+			color: $bio-white;
+			
+			&:hover{
+				background: $bio-white;
+				color: $bio-orange;
+			}
+		}
+
+		&.compact {
+			min-width: 130px;
+			
+			@include breakpoint($medium){
+				min-width: 160px;
+			}
 		}
 
 		&.whiteText{
@@ -44,6 +71,22 @@
 				color: $bio-black;
 				border: 2px solid $bio-orange;
             }
+		}
+		
+        &.long {
+			padding: 10px 15px;
+		}
+		
+		&.mediumLong{
+			width: 200px;
+		}
+
+		&.alignCenter {
+			margin: 0 auto;
+		}
+
+		&.alignRight{
+			margin: 0 0 0 auto;
 		}
 	}
 }
